refresh

+++ print_lowhexadecimal.c
@@ -7,9 +7,9 @@
 int print_lowhexadecimal(va_list args)
 {
 	char *str;
-	unsigned long int n = 0, number = 0, j = 0, aux = 0, i = number, count = 0;
+	unsigned int n = 0, number = 0, j = 0, aux = 0, i = number, count = 0;
 
-	number = va_arg(args, unsigned long int);
+	number = va_arg(args, unsigned int);
 	if (number == 0)
 	{
 		_putchar('0');
